refactor: 컴포넌트 분리 및 파일 확장자 변경
- jsx 파일 tsx로 변경
- ProductList 컴포넌트 분리
