+++ packages/assignment-5/src/refactoring/components/ProductItem.tsx
