+++ packages/assignment-5/src/refactoring/components/ProductSearch.tsx
